refactor(arch): remove proxy/launch layering inversions (resolve, mode, auth)
Three low-effort moves so the data-plane layers stop depending "up" on
cli/daemon internals:

- F2: split the clap/IPC/CliExit-free catalog matcher (CatalogRow,
  ResolveError, resolve_model_with_candidates) into launch/resolve.rs;
  cli/resolve.rs re-exports it and keeps the IPC fetch + CliExit
  wrappers. Proxy now imports the matcher from launch, not cli.
- F4: LaunchMode::resolve takes the domain enum; the clap->domain
  From conversion moves to cli/cli_args.rs so launch no longer imports
  cli::cli_args.
- F6: constant_time_eq + extract_bearer move to util/http_auth.rs;
  daemon and proxy auth both depend down on one copy.

grep -rn 'crate::cli::{resolve,exit_codes,cli_args::LaunchMode}'
src/proxy src/launch is now empty. Wire shapes and exit codes
unchanged; full suite green.
