+++ src/util/mod.rs
@@ -4,6 +4,7 @@ pub mod config_patch;
 pub mod datetime;
 pub mod file_security;
 pub mod hex;
+pub mod http_auth;
 pub mod logging;
 pub mod model_caches;
 pub mod paths;
